refactor(node-ui): decompose ConvictionDetailView into sections + shared hooks (#1348, #1375)
Behavior-preserving decomposition of the PCA detail (Manage) view — no behavior,
copy, testid, or gating change.

- #1348: split the ~290-line DetailBody into FundingSection / PublishingWalletsSection /
  LifecycleSection; ConvictionDetailView.tsx 879 -> 283 lines. Owner-gating stays
  centralized on one ownerWritesEnabled + ownerTitle threaded to each section; the
  wallet probe stays ungated. Removed dead code (ownerIsPrimary, the dead
  ownerOnlyReason variant, the unused onSwitchNetwork prop).
- #1375 item 4: extract useWalletTxProgress — the device-confirm progress reducer
  duplicated near-verbatim between CreatePcaModal and DetailBody — into one shared
  hook (optional-approve .find form); adopted in both, visible labels preserved.
- #1375 item 5: extract useConnectedWalletPcas ({ids,loading,error,refresh}) — the
  ERC721Enumerable connected-wallet discovery — out of ConvictionOverview.
- #1375 item 1: extract useWalletBootstrap, usePendingTopUpReconciliation,
  useRenewalChaining + pcaAction/detailWalletTx helpers.

Deferred to their own PRs: #1350 (ApproveWalletsModal reducer), #1375 item 2
(owner-access model), #1375 item 3 (resolve-signer-once). Tracked in #1375.

Preserved invariants: never-false-green probe; #9 registered-vs-submitted honesty;
the runFund error-ladder order + warning tri-state; inv-16 wallet-prompt safety
(onProgress wired only in wallet mode); inv-17 owner precedence; refresh() topology
(renewal deliberately omits refresh — it mints a replacement account); renewal
deregisterFrom + seedAgentsResolved; snapshot/agents stale-guards.

Verified: 445 PCA tests green (incl. a new non-owner gating-matrix test covering
top-up/approve/renew disabled + probe enabled across all three sections); tsc zero
errors in the changed areas; build:ui exit 0; 5-lens adversarial gate CLEAN
(0 must-fix) — runFund catch byte-identical, useWalletTxProgress behavior-identical,
gating centralized, no refresh/testid/hooks-rule drift.

Fixes #1348
