+++ packages/node-ui/src/ui/pages/conviction/CreatePcaModal.tsx
@@ -1,4 +1,4 @@
-import React, { useCallback, useMemo, useState } from 'react';
+import React, { useMemo, useState } from 'react';
 import { useFetch } from '../../hooks.js';
 import {
   fetchWalletsBalances,
@@ -21,7 +21,7 @@ import {
   WalletReceiptRevertedError,
   WalletReceiptWaitError,
 } from '../../web3/walletTxError.js';
-import type { WalletTxProgressEvent } from '../../web3/walletOwnerActionSubmitter.js';
+import { useWalletTxProgress } from '../../pca/useWalletTxProgress.js';
 import {
   DeviceConfirmProgress,
   DiscountTierLadder,
@@ -47,7 +47,8 @@ const AMOUNT_RE = /^\d+(\.\d+)?$/;
 function initialHardwareCreateSteps(): DeviceConfirmStep[] {
   return [
     { id: 'approve', label: 'Approve exact TRAC allowance', state: 'pending' },
-    { id: 'create', label: 'Sign Create PCA', state: 'pending' },
+    // id normalized to the shared 'action' step (label unchanged); the id is non-visible.
+    { id: 'action', label: 'Sign Create PCA', state: 'pending' },
     { id: 'confirm', label: 'Confirm on-chain receipt', state: 'pending' },
   ];
 }
@@ -114,65 +115,19 @@ export function CreatePcaModal({
   const walletChainId = useWalletStore((s) => s.chainId);
   const walletWrongNetwork = useWalletStore((s) => isWrongNetwork(s));
   const hardwareSelected = ownerKey === 'hardware';
-  const [hardwareCreateSteps, setHardwareCreateSteps] = useState<DeviceConfirmStep[]>([]);
-  const [hardwareCreateLabel, setHardwareCreateLabel] = useState<string>('Confirm on your device');
-  const onWalletProgress = useCallback((event: WalletTxProgressEvent) => {
-    setHardwareCreateLabel(() => {
-      if (event.step === 'approve') {
-        if (event.state === 'skipped' || event.state === 'confirmed') return 'Allowance ready — continue to Create PCA';
-        if (event.state === 'failed') return 'Wallet transaction failed';
-        return 'Confirm on your device (1 of 2): approve TRAC';
-      }
-      if (event.state === 'active') return 'Confirm on your device (2 of 2): Create PCA';
-      if (event.state === 'submitted') return 'Waiting for on-chain confirmation';
-      if (event.state === 'confirmed') return 'Create confirmed on-chain';
-      return 'Wallet transaction failed';
-    });
-    setHardwareCreateSteps((prev) => {
-      const next = (prev.length ? prev : initialHardwareCreateSteps()).map((s) => ({ ...s }));
-      const approve = next[0]!;
-      const create = next[1]!;
-      const confirm = next[2]!;
-      if (event.step === 'approve') {
-        if (event.state === 'skipped') {
-          approve.state = 'confirmed';
-          approve.label = 'TRAC allowance already sufficient';
-        } else if (event.state === 'active' || event.state === 'submitted') {
-          approve.state = 'active';
-          approve.txHash = event.txHash;
-        } else if (event.state === 'confirmed') {
-          approve.state = 'confirmed';
-          approve.txHash = event.txHash;
-        } else if (event.state === 'failed') {
-          approve.state = 'failed';
-          approve.txHash = event.txHash;
-          approve.error = describeWalletTxError(event.error, 'approve').message;
-        }
-      } else {
-        if (event.state === 'active') {
-          if (approve.state === 'pending') approve.state = 'confirmed';
-          create.state = 'active';
-        } else if (event.state === 'submitted') {
-          if (approve.state === 'pending') approve.state = 'confirmed';
-          create.state = 'confirmed';
-          create.txHash = event.txHash;
-          confirm.state = 'active';
-        } else if (event.state === 'confirmed') {
-          if (approve.state === 'pending') approve.state = 'confirmed';
-          create.state = 'confirmed';
-          create.txHash = event.txHash;
-          confirm.state = 'confirmed';
-          confirm.txHash = event.txHash;
-        } else if (event.state === 'failed') {
-          create.state = 'failed';
-          create.txHash = event.txHash;
-          create.error = describeWalletTxError(event.error, 'action').message;
-        }
-      }
-      return next;
-    });
-  }, []);
-  const owner = useOwnerActionSubmitter({ ownerKey, onWalletProgress });
+  const createProgress = useWalletTxProgress({
+    labels: {
+      idle: 'Confirm on your device',
+      approveActive: 'Confirm on your device (1 of 2): approve TRAC',
+      approveReady: 'Allowance ready — continue to Create PCA',
+      actionActive: 'Confirm on your device (2 of 2): Create PCA',
+      submitted: 'Waiting for on-chain confirmation',
+      confirmed: 'Create confirmed on-chain',
+      failed: 'Wallet transaction failed',
+    },
+    initialSteps: initialHardwareCreateSteps,
+  });
+  const owner = useOwnerActionSubmitter({ ownerKey, onWalletProgress: createProgress.onProgress });
   const finishCreate = usePcaStore((s) => s.finishCreate);
   const setCreatePending = usePcaStore((s) => s.setCreatePending);
   const clearCreatePending = usePcaStore((s) => s.clearCreatePending);
@@ -281,11 +236,9 @@ export function CreatePcaModal({
     setPhase('creating');
     setError(null);
     if (hardwareSelected) {
-      setHardwareCreateSteps(initialHardwareCreateSteps());
-      setHardwareCreateLabel('Confirm on your device (1 of 2): approve TRAC');
+      createProgress.begin('create');
     } else {
-      setHardwareCreateSteps([]);
-      setHardwareCreateLabel('Confirm on your device');
+      createProgress.reset();
     }
     // Set the double-mint guard at SUBMIT — before the await — so an ambiguous
     // failure (network drop / 500 / timeout, or the browser closing) AFTER the
@@ -534,8 +487,8 @@ export function CreatePcaModal({
         )}
         {hardwareSelected && phase === 'creating' && (
           <DeviceConfirmProgress
-            steps={hardwareCreateSteps.length ? hardwareCreateSteps : initialHardwareCreateSteps()}
-            currentLabel={hardwareCreateLabel}
+            steps={createProgress.steps.length ? createProgress.steps : initialHardwareCreateSteps()}
+            currentLabel={createProgress.currentLabel}
             blockExplorerUrl={explorer}
           />
         )}
